refactor: place AppState at app root, not in app-shell
- Move AppState to src/appState.ts (app-level, not feature-owned)
- Harden createAppState with per-field nullish coalescing
- Stop exporting AppState helpers from app-shell barrel
- Keep app-shell/actions importing the type without depending on App.tsx
